move load_picam_png to io_py
